requirements.txt
flask
flask-migrate
flask-sqlalchemy
sqlalchemy-utils
psycopg2  # Postgres backend
